Add DBUS_INT64_MODIFIER define, turn on -Wformat
https://bugs.freedesktop.org/show_bug.cgi?id=19195

We were previously using -Wno-format because we didn't have
a #define for DBUS_INT64_MODIFIER, which was really lame because
it easily hid problems.

For now, just define it if we're on glibc; this is obviously
not strictly correct but it's safe, because the formatting
is only used in DBUS_VERBOSE mode, and in tools/dbus-monitor.
Ideally we get the the glib code relicensed.

+++ dbus/dbus-credentials.c
@@ -519,13 +519,13 @@ _dbus_credentials_to_string_append (DBusCredentials    *credentials,
   join = FALSE;
   if (credentials->unix_uid != DBUS_UID_UNSET)
     {
-      if (!_dbus_string_append_printf (string, "uid=%d", credentials->unix_uid))
+      if (!_dbus_string_append_printf (string, "uid=%d", (int) credentials->unix_uid))
         goto oom;
       join = TRUE;
     }
   if (credentials->unix_pid != DBUS_PID_UNSET)
     {
-      if (!_dbus_string_append_printf (string, "%spid=%d", join ? " " : "", credentials->unix_pid))
+      if (!_dbus_string_append_printf (string, "%spid=%d", join ? " " : "", (int) credentials->unix_pid))
         goto oom;
       join = TRUE;
     }
